refactor: split up type helper
