+++ src/core/util/clone.ts
@@ -1,0 +1,5 @@
+import { cloneDeep } from 'lodash';
+
+export function clone(value: any): any {
+  return cloneDeep(value);
+}
