+++ src/core/util/index.ts
@@ -1,6 +1,9 @@
+export * from './clone';
 export * from './create-builder.fn';
 export * from './create-processor.fn';
+export * from './is-array-of-buildables';
+export * from './is-buildable';
 export * from './is-builder.fn';
+export * from './is-defined';
 export * from './is-processor.fn';
-export * from './type.helper';
 export * from './unwrap-if-buildable';
