+++ src/core/util/is-array-of-buildables.ts
@@ -1,0 +1,11 @@
+import { Buildable } from '../types';
+import { isBuildable } from './is-buildable';
+import { isDefined } from './is-defined';
+
+export function isArrayOfBuildables(value: any): value is Buildable[] {
+  if (!isDefined(value) || !Array.isArray(value)) {
+    return false;
+  }
+
+  return value.length > 0 && value.every(v => isBuildable(v));
+}
