+++ src/core/util/is-defined.ts
@@ -1,0 +1,3 @@
+export function isDefined<T>(value: T | undefined | null): value is T {
+  return value !== undefined && value !== null;
+}
